requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai==1.43.0
langchain_community==0.2.15
supabase==2.7.3
tiktoken==0.7.0
langchain-openai
httpx[http2]